logically moved a majority of simulation functions to game_state.py as functions/methods; logically moved generate hand functions from simulation.py to cards.py; reorganized tests to support new code structure; new csvs and chart generated from 50_000 sim; reduced the amount of GameState objects created and destroyed per simulation case
