+++ src/blackjack/data/data_split.csv
@@ -1,101 +1,101 @@
 player hand total,player hand texture,dealer face up,player choice,expected value
-20,split,11,split,-21.36
-18,split,11,split,-12.27
-16,split,11,split,-16.45
-14,split,11,split,-19.91
-12,split,11,split,-20.39
-10,split,11,split,-20.75
-8,split,11,split,-18.58
-6,split,11,split,-17.39
-4,split,11,split,-16.88
-2,split,11,split,-5.73
-20,split,10,split,-18.39
-18,split,10,split,-9.05
-16,split,10,split,-13.17
-14,split,10,split,-16.86
-12,split,10,split,-18.06
-10,split,10,split,-19.02
-8,split,10,split,-16.32
-6,split,10,split,-14.59
-4,split,10,split,-13.25
-2,split,10,split,2.44
+20,split,11,split,-20.83
+18,split,11,split,-12.12
+16,split,11,split,-16.81
+14,split,11,split,-20.19
+12,split,11,split,-20.28
+10,split,11,split,-20.62
+8,split,11,split,-18.8
+6,split,11,split,-17.75
+4,split,11,split,-16.84
+2,split,11,split,-5.68
+20,split,10,split,-19.08
+18,split,10,split,-9.02
+16,split,10,split,-12.96
+14,split,10,split,-17.39
+12,split,10,split,-18.43
+10,split,10,split,-18.94
+8,split,10,split,-16.43
+6,split,10,split,-14.54
+4,split,10,split,-13.48
+2,split,10,split,2.42
 20,split,9,split,-14.72
-18,split,9,split,-1.96
-16,split,9,split,-9.68
-14,split,9,split,-14.11
-12,split,9,split,-14.91
-10,split,9,split,-15.49
-8,split,9,split,-12.68
-6,split,9,split,-10.72
-4,split,9,split,-9.43
-2,split,9,split,5.78
-20,split,8,split,-5.8
-18,split,8,split,5.71
-16,split,8,split,-0.43
-14,split,8,split,-10.16
-12,split,8,split,-10.7
-10,split,8,split,-11.51
-8,split,8,split,-7.96
-6,split,8,split,-5.38
-4,split,8,split,-4.17
-2,split,8,split,8.78
-20,split,7,split,2.77
-18,split,7,split,8.86
-16,split,7,split,8.15
-14,split,7,split,-1.15
-12,split,7,split,-5.89
-10,split,7,split,-7.08
-8,split,7,split,-3.95
-6,split,7,split,-0.77
-4,split,7,split,0.97
-2,split,7,split,11.76
-20,split,6,split,8.8
-18,split,6,split,11.55
+18,split,9,split,-1.95
+16,split,9,split,-9.91
+14,split,9,split,-14.17
+12,split,9,split,-14.92
+10,split,9,split,-15.9
+8,split,9,split,-12.35
+6,split,9,split,-10.19
+4,split,9,split,-9.82
+2,split,9,split,6.0
+20,split,8,split,-5.46
+18,split,8,split,5.92
+16,split,8,split,-0.54
+14,split,8,split,-9.45
+12,split,8,split,-10.39
+10,split,8,split,-11.38
+8,split,8,split,-8.04
+6,split,8,split,-5.72
+4,split,8,split,-3.62
+2,split,8,split,8.65
+20,split,7,split,3.04
+18,split,7,split,8.76
+16,split,7,split,8.36
+14,split,7,split,-1.17
+12,split,7,split,-6.3
+10,split,7,split,-7.12
+8,split,7,split,-4.09
+6,split,7,split,-1.42
+4,split,7,split,0.21
+2,split,7,split,11.62
+20,split,6,split,9.35
+18,split,6,split,11.61
 16,split,6,split,9.81
-14,split,6,split,6.37
-12,split,6,split,5.23
-10,split,6,split,2.43
-8,split,6,split,5.69
-6,split,6,split,6.75
-4,split,6,split,7.82
-2,split,6,split,16.91
-20,split,5,split,6.51
-18,split,5,split,9.61
-16,split,5,split,7.92
-14,split,5,split,4.13
-12,split,5,split,2.84
-10,split,5,split,0.52
-8,split,5,split,3.45
-6,split,5,split,4.07
-4,split,5,split,5.03
-2,split,5,split,15.75
-20,split,4,split,2.9
-18,split,4,split,8.15
-16,split,4,split,5.95
-14,split,4,split,1.32
-12,split,4,split,0.69
-10,split,4,split,-2.08
-8,split,4,split,0.69
-6,split,4,split,1.79
-4,split,4,split,3.1
-2,split,4,split,14.15
-20,split,3,split,0.86
-18,split,3,split,6.95
-16,split,3,split,3.97
-14,split,3,split,-1.0
-12,split,3,split,-2.22
-10,split,3,split,-4.94
-8,split,3,split,-1.89
-6,split,3,split,-0.33
-4,split,3,split,0.61
-2,split,3,split,13.22
-20,split,2,split,-2.49
-18,split,2,split,4.47
-16,split,2,split,2.15
-14,split,2,split,-3.33
-12,split,2,split,-4.54
-10,split,2,split,-6.18
-8,split,2,split,-4.54
-6,split,2,split,-2.78
-4,split,2,split,-1.17
-2,split,2,split,12.01
+14,split,6,split,6.06
+12,split,6,split,5.02
+10,split,6,split,2.78
+8,split,6,split,5.9
+6,split,6,split,6.9
+4,split,6,split,7.74
+2,split,6,split,17.04
+20,split,5,split,6.39
+18,split,5,split,9.68
+16,split,5,split,7.32
+14,split,5,split,3.85
+12,split,5,split,2.81
+10,split,5,split,0.41
+8,split,5,split,2.74
+6,split,5,split,4.48
+4,split,5,split,5.22
+2,split,5,split,15.74
+20,split,4,split,3.67
+18,split,4,split,8.04
+16,split,4,split,5.55
+14,split,4,split,1.94
+12,split,4,split,0.05
+10,split,4,split,-2.26
+8,split,4,split,0.68
+6,split,4,split,2.25
+4,split,4,split,2.59
+2,split,4,split,14.61
+20,split,3,split,0.2
+18,split,3,split,6.24
+16,split,3,split,3.96
+14,split,3,split,-0.46
+12,split,3,split,-1.72
+10,split,3,split,-4.4
+8,split,3,split,-2.08
+6,split,3,split,-0.29
+4,split,3,split,0.83
+2,split,3,split,13.39
+20,split,2,split,-1.93
+18,split,2,split,4.62
+16,split,2,split,1.87
+14,split,2,split,-2.61
+12,split,2,split,-4.29
+10,split,2,split,-6.47
+8,split,2,split,-4.04
+6,split,2,split,-3.23
+4,split,2,split,-1.61
+2,split,2,split,12.19
