+++ src/blackjack/helper/cards.py
@@ -1,3 +1,5 @@
+import random
+
 class Card:
     def __init__(self, number: int, suit: str) -> None:
         if number < 1:
@@ -69,6 +71,50 @@ def add(hand: list[Card]) -> int:
             continue
         else:
             return total
-
-
     
+
+# NOTE: *sigh* python 3.10 added match-case, but is it worth the rebase?
+def generate_hand(hand_total: int, hand_texture: str) -> Hand:
+    suits = ('heart','diamond','club','spade')
+
+    if hand_total < 12 and hand_texture == 'soft':
+        raise ValueError("cannot create a soft hand with a value of less than 12")
+
+    # split just make sure that both cards are the same when splitting, i.e. 5,5 for a requested hand total of 10
+    elif hand_texture == 'split':
+        first_card = int(hand_total/2)
+        second_card = int(hand_total/2)
+
+        hand = Hand()
+        hand.hand_list = [Card(first_card, random.choice(suits)), Card(second_card, random.choice(suits))]
+        return hand
+    
+    elif hand_texture == 'hard':
+        minimum_int = hand_total - 10
+
+        # this line is an algorithm that makes it so that any possible combination cards could be generated
+        first_card = random.randint(max(1,minimum_int), min(10, hand_total-1))
+        second_card = hand_total - first_card
+
+        hand = Hand()
+        hand.hand_list = [Card(first_card,random.choice(suits)), Card(second_card,random.choice(suits))]
+        return hand
+    
+    elif hand_texture == 'soft':
+        first_card = 11
+        second_card = hand_total - first_card
+
+        hand = Hand()
+        hand.hand_list = [Card(first_card,random.choice(suits)), Card(second_card,random.choice(suits))]
+        return hand
+    
+    else:
+        raise Exception("Unexpected Error within generate_hand()")
+
+
+def generate_dealer_hand(face_up_card: int) -> Hand:
+    suits = ('heart','diamond','club','spade')
+    simulated_deck = [2,3,4,5,6,7,8,9,10,10,10,10,11]
+    hand = Hand()
+    hand.hand_list = [Card(face_up_card, random.choice(suits)), Card(random.choice(simulated_deck), random.choice(suits))]
+    return hand
